+++ polars_utils/validate.py
@@ -1,0 +1,98 @@
+import json
+from pathlib import Path
+
+import boto3
+import pointblank as pb
+import polars as pl
+import polars.selectors as cs
+
+from polars_utils import utils
+from polars_utils.logger import get_logger
+
+logger = get_logger(__name__)
+
+
+def read_parquet(
+    source: str | Path, exclude_complex_types: bool = False
+) -> pl.DataFrame:
+    """Reads in a parquet in a format suitable for validating
+
+    Args:
+        source (str | Path): the full path in s3 of the dataset to be validated
+        exclude_complex_types (bool, optional): whether or not to exclude types which cannot be
+            validated using pointblank (ie., Structs, Lists or similar). Defaults to False.
+
+    Returns:
+        pl.DataFrame: the raw data as a polars Dataframe
+    """
+    raw = pl.scan_parquet(
+        source,
+        cast_options=pl.ScanCastOptions(missing_struct_fields="insert"),
+        extra_columns="ignore",
+    )
+    if not exclude_complex_types:
+        return raw.collect()
+
+    return raw.select(~cs.by_dtype(pl.Struct, pl.List)).collect()
+
+
+def write_reports(validation: pb.Validate, bucket_name: str, reports_path: str) -> None:
+    """Writes the reports for a given `pb.Validate` object to s3, including:
+        - summary report as an HTML `pb.GT`
+        - a pl.Dataframe for each failed step (if any) including every failing record
+
+    Args:
+        validation (pb.Validate): the result of interrogating the defined validation
+        bucket_name (str): the bucket to save reports to
+            - shoud correspond to workspace / feature branch name
+        reports_path (str): the filepath for the reports
+
+    Raises:
+        AssertionError: in case of the dataset failing the validation rules
+    """
+
+    report = validation.get_tabular_report()
+
+    s3_client = boto3.client("s3")
+    s3_client.put_object(
+        Body=report.as_raw_html(inline_css=True, make_page=True),
+        Bucket=bucket_name,
+        Key=f"{reports_path}/index.html",
+    )
+    try:
+        validation.assert_below_threshold(level="warning")
+    except AssertionError:
+        logger.error("Data validation failed. See report for details.")
+        steps = json.loads(validation.get_json_report())
+        # JSON report includes a detailed list of each validation step, including failures
+        # Note that some 'steps' result in several steps in the execution
+        # eg. a null check over several columns
+        for step in steps:
+            _report_on_fail(step, validation, bucket_name, reports_path)
+        raise  # ensures that the task fails if any warnings / errors
+
+
+def _report_on_fail(
+    step: dict, validation: pb.Validate, bucket_name: str, path: str
+) -> None:
+    """Checks a given pb.Validate step for failures and writes the failed records to S3 if present.
+
+    Args:
+        step (dict): metadata on the validation step result
+        validation (pb.Validate): the Validate object containing the resulting data
+        bucket_name (str): the bucket in which to write reports
+        path (str): the filepath in the bucket to write, should include the validation report name
+    """
+    if step["all_passed"]:
+        return
+
+    step_idx = step["i"]
+    assertion = step["assertion_type"]
+    _col_or_cols = step["column"]  # could be a string or a list
+    columns = "_".join(_col_or_cols) if isinstance(_col_or_cols, list) else _col_or_cols
+
+    failed_records_df = validation.get_data_extracts(step_idx, frame=True)
+    utils.write_to_parquet(
+        failed_records_df,  # type: ignore = frame=True returns a df
+        f"s3://{bucket_name}/{path}/failed_step_{step_idx}_{assertion}_{columns}.parquet",
+    )